o.e.l.systemtap.ui.dashboard cleanup.
* fix the activator stop to not throw NPE on stop due to non-existing
files
* remove excessive logging
* use system cursors
* delete commented code
* remove listeners that don't do anything
* use adapters instead of listeners

Reviewed-on: https://git.eclipse.org/r/9561
Tested-by: Hudson CI
IP-Clean: Camilo Bernal <cabernal@redhat.com>
Tested-by: Camilo Bernal <cabernal@redhat.com>

+++ systemtap/org.eclipse.linuxtools.systemtap.ui.dashboard/src/org/eclipse/linuxtools/systemtap/ui/dashboard/internal/DashboardPlugin.java
@@ -57,14 +57,18 @@ public class DashboardPlugin extends AbstractUIPlugin {
 	}
 
 	private void removeFolder(File folder) {
-		File[] files = folder.listFiles();
-		for(int j=0; j<files.length; j++) {
-			if(files[j].isDirectory())
-				removeFolder(files[j]);
-			else
-				files[j].delete();
+		if (folder != null) {
+			File[] files = folder.listFiles();
+			if (files != null) {
+				for (File file : folder.listFiles()) {
+					if (file.isDirectory())
+						removeFolder(file);
+					else
+						file.delete();
+				}
+			}
+			folder.delete();
 		}
-		folder.delete();
 	}
 
 	/**
